pyproject.toml
[project]
name = "model-runtime-authority"
version = "0.1.0"
description = "MandateCore authority validation with a Streamlit workbench for decision review."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11,<3",
    "pyyaml>=6.0,<7",
    "streamlit>=1.47,<2",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115,<1",
    "httpx>=0.28,<1",
    "uvicorn>=0.35,<1",
]
